The worker is split into several classes
